pyproject.toml
[tool.poetry]
name = "rag-cli"
version = "0.1.0"
description = "Retrieval Augmented Generation CLI Tool"
authors = ["jacobh310 <jacob.hernandez0830@gmail.com>"]
readme = "README.md"
packages = [{include = "rag", from = "src"}]

[tool.poetry.scripts]
rag = "rag.entrypoint:cli"

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
pandas = "2.2.2"
numpy = "1.26.4"
pyyaml = "6.0.1"
sec-edgar-downloader = "5.0.2"
boto3 = "1.34.108"
langchain = "0.0.240"
click = "8.1.8"
python-dotenv = "1.0.1"
sagemaker = "2.220.0"
pinecone = "3.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

